NotificationManagerService: Fix notification led bug (Part 1/2)
We can assume that if the user turned the screen off while there was
still an active notification then they wanted to keep the notification
for later. In this case we shouldn't flash the notification light.

The problem for some special notifications (such as missed calls)
is that the screen is automatically turned on, and then when it is
turned back off the led does not get enabled. For these notifications
we use a flag to force the notification light even if the screen was
turned off.

 This is the 2nd commit message:

fw: hide FLAG_FORCE_LED_SCREEN_OFF

 This is the 3rd commit message:

fw: Fix notification service

Change-Id: I3eeabc3b63b44bc92ae7ce5d6253e312e6510d45